refactor(core): 收拢 Natal 生成路径
