+++ crates/ziwei_core/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ziwei_core"
 version = "0.1.0"
-description = "Core domain model and chart construction for Ziwei"
+description = "Core domain model for immutable Ziwei natal charts"
 publish = false
 
 edition.workspace = true
